Add multi-tenant support (#24)
This change adds a `multitenant_mode` setting that will allow migrating resources owned by other projects. It will be enabled by default, requiring admin privileges.

If enabled, the owner project/user will be reported as an associated resource (dependency), being migrated and referenced during the migrations.

The tests are also updated to cover multi-tenant mode. A set of temporary credentials will be created for every test module. If multi-tenant mode is enabled, the tests will use one tenant for creating the test resources (the resource owner) and a separate tenant for initiating the migration (called "requester").